# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kartly REST API
  version: '1.0'
  description: |
    JSON API behind the Kartly storefront. The server-rendered pages and this API run the
    same service layer, so anything the website can do is available here.

    **Authentication** is a session cookie (`kartly.sid`), set by `POST /auth/login` or
    `POST /auth/register` and sent automatically by browsers.

    **CSRF:** every unsafe request (POST, PATCH, DELETE) must carry the session's token in an
    `X-CSRF-Token` header. Fetch one from `GET /auth/csrf-token`. The token is regenerated
    when the session is (that is, on sign-in and sign-out), so read it again from the login
    response.

    **Guest carts:** cart endpoints work without signing in. The cart is tied to the session
    and is merged into the user's cart on the next sign-in or registration.

    **Errors** always use the shape `{"error": {"code", "message", "details"?}}`.

    **Rate limits:** 900 requests per 15 minutes per IP for the API as a whole, and 20 per
    15 minutes for `/auth/login` and `/auth/register`.
  license:
    name: MIT
    url: https://github.com/VivekReddy2001/shopping_cart/blob/main/LICENSE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://{host}/api/v1
    description: Deployed instance
    variables:
      host:
        default: kartly.onrender.com
...
paths:
...
  /products:
    get:
      tags: [Catalogue]
      summary: Search, filter and page the catalogue
      description: |
        `q` accepts natural language: `phones under 20k`, `samsung between 10000 and 25000`.
        Price phrases are lifted out of the text and applied as filters. Invalid query values
        fall back to their defaults instead of returning an error.
      security: []
      parameters:
        - {
            name: q,
            in: query,
            schema: { type: string, maxLength: 100 },
            example: phones under 20k,
          }
        - { name: category, in: query, schema: { type: string }, description: Category slug }
        - {
            name: brand,
            in: query,
            schema: { type: array, items: { type: string } },
            style: form,
            explode: true,
          }
        - { name: minPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: rating, in: query, schema: { type: number, minimum: 0, maximum: 5 } }
        - { name: inStock, in: query, schema: { type: boolean } }
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, newest, rating, discount]
            default: relevance
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 60, default: 24 },
          }
      responses:
        '200':
          description: A page of products plus facet counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      pages: { type: integer }
                      limit: { type: integer }
                      appliedFilters: { type: object }
                      facets: { type: object }
    post:
      tags: [Admin]
      summary: Create a product
      description: Admin only. `multipart/form-data`, because it carries the product image.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, description, price, stock, category, brand, image]
              properties:
                name: { type: string }
                description: { type: string }
                price: { type: number }
                mrp: { type: number }
                stock: { type: integer }
                category: { type: string, description: Category id }
                brand: { type: string, description: Brand id }
                tags: { type: string, description: Comma-separated }
                featured: { type: boolean }
                isActive: { type: boolean, default: true }
                image:
                  type: string
                  format: binary
                  description: JPEG, PNG, WebP or GIF, 2 MB max. Verified by magic bytes.
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProductDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /products/{idOrSlug}:
    get:
      tags: [Catalogue]
      summary: One product by id or slug
      security: []
      parameters:
        - {
            name: idOrSlug,
            in: path,
            required: true,
            schema: { type: string },
            example: apple-iphone-11,
          }
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProductDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{id}:
    patch:
      tags: [Admin]
      summary: Update a product
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Any subset of the create fields. Omitted fields are left untouched.
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProductDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Archive a product
      description: |
        Products are never hard-deleted — `isActive` is set to `false` so past orders,
        reviews and analytics stay valid. The product disappears from the storefront.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Archived }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kartly.sid
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        url: { type: string }
        image: { type: string }
        brand: { $ref: '#/components/schemas/Ref' }
        category: { $ref: '#/components/schemas/Ref' }
        price: { type: number }
        mrp: { type: number }
        discountPercent: { type: integer }
        stock: { type: integer }
        inStock: { type: boolean }
        rating:
          type: object
          properties:
            average: { type: number }
            count: { type: integer }
        featured: { type: boolean }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description: { type: string }
            highlights: { type: array, items: { type: string } }
            tags: { type: array, items: { type: string } }
            isActive: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    Ref:
      type: object
      nullable: true
      properties:
        name: { type: string }
        slug: { type: string }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string }
            details:
              type: object
              additionalProperties: { type: string }
              description: Field-level messages, for validation failures.
  responses:
...
    Forbidden:
      description: Signed in, but not allowed (including the read-only demo admin)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: The body failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - sessionCookie: []
    csrfToken: []